government-scheme-agent/requirements.txt
streamlit
langchain
langchain-openai
python-dotenv
pydantic
pandas
ddgs
tiktoken
